+++ sprockets/http/runner.py
@@ -1,18 +1,20 @@
 """
 Run a Tornado HTTP service.
 
-- :func:`.run`: calls a ``make_app`` *callable*, configures the
-  environment intelligently, and runs the application.
 - :class:`.Runner`: encapsulates the running of the application
+- :class:`.RunCommand`: distutils command to runs an application
 
 """
+from distutils import cmd, errors, log
 import logging
+import os.path
 import signal
 import sys
 
 from tornado import concurrent, httpserver, ioloop
 
 import sprockets.logging
+import sprockets.http
 
 
 class Runner(object):
@@ -192,3 +194,80 @@ class Runner(object):
         # If no futures were return, invoke on shutdown ready
         if not self._pending_callbacks:
             self._on_shutdown_ready()
+
+
+class RunCommand(cmd.Command):
+    """
+    Simple distutils.Command that calls :func:`sprockets.http.run`
+
+    This is installed as the httprun distutils command when you
+    install the ``sprockets.http`` module.
+
+    """
+
+    description = 'Run a sprockets.http application.'
+    user_options = [
+        ('application=', 'a',
+         'application callable in `pkg.mod:func` syntax'),
+        ('env-file=', 'e', 'environment file to import'),
+        ('port=', 'p', 'port for the application to listen on'),
+    ]
+
+    def initialize_options(self):
+        self.application = None
+        self.env_file = None
+        self.port = None
+
+    def finalize_options(self):
+        if not self.application:
+            raise errors.DistutilsArgError('application is required')
+        if self.env_file and not os.path.exists(self.env_file):
+            raise errors.DistutilsArgError(
+                'environment file "{}" does not exist'.format(
+                    self.env_file))
+
+    def run(self):
+        self._read_environment()
+        if self.port:
+            log.info('overriding port to %s', self.port)
+            os.environ['PORT'] = self.port
+        app_factory = self._find_callable()
+        if self.dry_run:
+            log.info('would run %r', app_factory)
+        else:
+            log.info('running %r', app_factory)
+            sprockets.http.run(app_factory)
+
+    def _read_environment(self):
+        if not self.env_file:
+            return
+
+        with open(self.env_file) as env_file:
+            for line in env_file.readlines():
+                orig_line = line.strip()
+                if '#' in line:
+                    line = line[:line.index('#')]
+                if line.startswith('export '):
+                    line = line[7:]
+
+                name, sep, value = line.strip().partition('=')
+                if sep == '=':
+                    if (value.startswith(('"', "'")) and
+                            value.endswith(value[0])):
+                        value = value[1:-1]
+                    if value:
+                        log.info('setting environment %s=%s', name, value)
+                        os.environ[name] = value
+                    else:
+                        log.info('removing %s from environment', name)
+                        os.environ.pop(name, None)
+                elif line:
+                    log.info('malformed environment line %r ignored',
+                             orig_line)
+
+    def _find_callable(self):
+        app_module, callable_name = self.application.split(':')
+        mod = __import__(app_module)
+        for next_mod in app_module.split('.')[1:]:
+            mod = getattr(mod, next_mod)
+        return getattr(mod, callable_name)